requirements-backend.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
langchain==0.3.18
langchain-core==0.3.45
langchain-openai==0.3.9
langgraph==0.2.70
openai==1.66.5
typing_extensions==4.12.2
